Move code to specific analysis configuration
