+++ python/l1THistos.py
@@ -12,7 +12,6 @@ import uproot as up
 from scipy.special import expit
 
 import python.boost_hist as bh
-from python import pf_regions
 
 stuff = []
 
@@ -1740,105 +1739,6 @@ class CalibrationHistos(BaseTuples):
         values_fill.append(reference.pt)
         self.t_values.Fill(array('f', values_fill))
 
-
-
-
-
-class CorrOccupancyHistos(BaseHistos):
-    class CorrBoardOccupancyHistos:
-        def __init__(self, name, board):
-            self.h_totOcc = bh.TH1F(
-                f'{name}_{board}totOcc', 
-                f'{board} total occupancy; {board} total occ.', 
-                500, 0, 500)
-            self.h_regOcc = bh.TH1F(
-                f'{name}_{board}regOcc', 
-                f'{board} reg occupancy; {board} reg. occ.', 
-                100, 0, 100)
-            self.h_maxOcc = bh.TH1F(
-                f'{name}_{board}maxOcc', 
-                f'{board} max occupancy; {board} max occ.', 
-                100, 0, 100)
-            self.h_maxMult = bh.TH1F(
-                f'{name}_{board}maxMult', 
-                f'{board} max multiplicity per cluster; {board} max mult. per cluster', 
-                100, 0, 100)
-            self.eta_regions_idx = pf_regions.regions[board]
-            self.max_count = None
-            self.tot_count = None
-            # self.max_mult_percluster = 0
-
-        def fillRegion(self, ieta, occupancy, multpercluster):
-            if(ieta in self.eta_regions_idx):
-                self.h_regOcc.fill(occupancy)
-                if self.max_count is None:
-                    # print(f'--------- REset Board: {self.eta_regions_idx}')
-                    self.max_count = occupancy
-                else:
-                    # print(f'Occ: {occupancy}')
-                    # print(f'Old Max: {self.max_count}')
-                    self.max_count = np.maximum(self.max_count, occupancy)
-                    # print(f'New Max: {self.max_count}')
-
-                if self.tot_count is None:
-                    self.tot_count = occupancy
-                else:
-                    self.tot_count = self.tot_count + occupancy
-                # if multpercluster > self.max_mult_percluster:
-                #     self.max_mult_percluster = multpercluster
-
-        def fillBoard(self):
-            self.h_maxOcc.fill(self.max_count)
-            self.h_totOcc.fill(self.tot_count)
-            # self.h_maxMult.Fill(self.max_mult_percluster)
-            self.max_count = None
-            self.tot_count = None
-            # self.max_mult_percluster = 0
-
-
-    def __init__(self, name, root_file=None, debug=False):
-        if not root_file:
-
-            # self.h_avgOcc = ROOT.TProfile2D(
-            #     f'{name}_avgOcc',
-            #     'region avg occ; #eta, #phi;',
-            #     pf_regions.regionizer.n_eta_regions(),
-            #     array('d', pf_regions.regionizer.eta_boundaries_fiducial_),
-            #     pf_regions.regionizer.n_phi_regions(),
-            #     array('d', pf_regions.regionizer.phi_boundaries_fiducial_))
-
-            self.board_histos = []
-            for board in ['ALL', 'BRL', 'HGCNoTk', 'HGC']:
-                bhs = CorrOccupancyHistos.CorrBoardOccupancyHistos(name, board)
-                setattr(self, f'h_{board}totOcc', bhs.h_totOcc)
-                setattr(self, f'h_{board}regOcc', bhs.h_regOcc)
-                setattr(self, f'h_{board}maxOcc', bhs.h_maxOcc)
-                setattr(self, f'h_{board}maxMult', bhs.h_maxMult)
-                self.board_histos.append(bhs)
-
-        BaseHistos.__init__(self, name, root_file, debug)
-
-    def fill(self, objects):
-        # print(objects.show())
-        # print(objects.type.show())
-
-        for ieta in range(pf_regions.regionizer.n_eta_regions()):
-            for iphi in range(pf_regions.regionizer.n_phi_regions()):
-                objs_in_region = objects[objects[f'eta_reg_{ieta}'] & objects[f'phi_reg_{iphi}']]
-                occupancy = ak.count(objs_in_region, axis=1)
-                nmatch_percluster = 0
-                # if 'clidx' in objs_in_region.columns and not objs_in_region.empty:
-                #     nmatch_percluster = objs_in_region.clidx.value_counts().iloc[0]
-                # occupancy = objs_in_region.shape[0]
-
-                # self.h_avgOcc.Fill(pf_regions.regionizer.eta_centers[ieta],
-                #                    pf_regions.regionizer.phi_centers[iphi],
-                #                    occupancy)
-                for bhs in self.board_histos:
-                    bhs.fillRegion(ieta, occupancy, nmatch_percluster)
-
-        for bhs in self.board_histos:
-            bhs.fillBoard()
 
 
 # for convenience we define some sets
